// src/generated/types/GroupCalculationResultDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type GroupCalculationResultDto = {
  /**
   * @type number
   */
  totalStudents: number
  /**
   * @type number
   */
  requiredSessions: number
  /**
   * @type array
   */
  groupDistribution: {
    /**
     * @type string | undefined
     */
    courseGroupId?: string | undefined
    /**
     * @type string | undefined
     */
    courseGroupName?: string | undefined
    /**
     * @type number | undefined
     */
    studentCount?: number | undefined
    /**
     * @type number | undefined
     */
    recommendedSessions?: number | undefined
    /**
     * @type number | undefined
     */
    maxStudentsPerSession?: number | undefined
  }[]
  /**
   * @type array
   */
  labAvailability: {
    /**
     * @type string | undefined
     */
    labId?: string | undefined
    /**
     * @type string | undefined
     */
    labName?: string | undefined
    /**
     * @type number | undefined
     */
    capacity?: number | undefined
    /**
     * @type number | undefined
     */
    availableSlots?: number | undefined
  }[]
}