// src/generated/types/CreateScheduleDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type CreateScheduleDto = {
  /**
   * @type string
   */
  courseGroupId: string
  /**
   * @type string | undefined
   */
  labId?: string | undefined
  /**
   * @type string, date-time
   */
  dateTime: Date
  /**
   * @type string
   */
  assistantId: string
  /**
   * @type number
   */
  maxStudents: number
}